I moved the API logic to "Date Options" + made a components folder and container folder
